# seq: use Floyd's combination algorithm to sample indices