… CHANGELOG.md
… src/seq/index.rs
… src/seq/mod.rs