Adapt #144 for sampling indices

This isn't the fully generic implementation that @burdges wrote; we could make this more generic. I'm guessing we could include an arbitrary lower bound without performance overhead (since the compiler can probably eliminate a constant of 0), and could also use generic typing.

I'm not really sure if it should be committed like this. It cleans up the code and improves performance in some cases, but isn't optimal in others. We could keep the old `sample_indices_inplace` and just replace `sample_indices_cache`.

```
# before
test gen_1k_sample_iter              ... bench:         390 ns/iter (+/- 4) = 2625 MB/s                                                
test misc_sample_indices_100_of_1k   ... bench:         806 ns/iter (+/- 8)                                                            
test misc_sample_indices_10_of_1k    ... bench:         594 ns/iter (+/- 9)                                                            
test misc_sample_indices_50_of_1k    ... bench:         472 ns/iter (+/- 8)                                                            
test misc_sample_iter_10_of_100      ... bench:       1,072 ns/iter (+/- 15)                                                           
test misc_sample_slice_10_of_100     ... bench:         160 ns/iter (+/- 2)                                                            
test misc_sample_slice_ref_10_of_100 ... bench:         160 ns/iter (+/- 2)                                                            
# after                                                                                                       
test gen_1k_sample_iter              ... bench:         391 ns/iter (+/- 27) = 2618 MB/s
test misc_sample_indices_100_of_1k   ... bench:       2,031 ns/iter (+/- 39)
test misc_sample_indices_10_of_1k    ... bench:          90 ns/iter (+/- 1)
test misc_sample_indices_50_of_1k    ... bench:         649 ns/iter (+/- 11)
test misc_sample_iter_10_of_100      ... bench:       1,083 ns/iter (+/- 16)
test misc_sample_slice_10_of_100     ... bench:         157 ns/iter (+/- 1)
test misc_sample_slice_ref_10_of_100 ... bench:         158 ns/iter (+/- 3)       
```

Edit: now:
```
test misc_sample_indices_100_of_1k           ... bench:         515 ns/iter (+/- 24)
test misc_sample_indices_10_of_1k            ... bench:          86 ns/iter (+/- 3)
test seq_iter_choose_multiple_10_of_100      ... bench:       1,137 ns/iter (+/- 319)
test seq_iter_choose_multiple_fill_10_of_100 ... bench:       1,088 ns/iter (+/- 35)
test seq_slice_choose_multiple_10_of_100     ... bench:         139 ns/iter (+/- 8)
````