@dhardy 2018-05-25
CC @vitiral who wrote the old implementation