@dhardy 2018-05-26
I realise that this is `O(m^2)` due to the use of `Vec::contains`, so for large `m=amount` it would make sense to use a tree or hash set instead, however for small `m` a simple `Vec` works very well.

The partial Fisher-Yates shuffle avoids the need for comparison and should thus be faster when the time/memory needed to allocate the whole sample space is not significant; essentially this algorithm is `O(n)` for `n=length`.

Does this mean that for efficiency we should have three backing implementations? Not ideal.