@dhardy 2018-06-01
Some more benchmarks:
```
test misc_sample_indices_cache_1000_of_100k   ... bench:      64,554 ns/iter (+/- 6,677)
test misc_sample_indices_cache_1000_of_10k    ... bench:      69,582 ns/iter (+/- 4,090)
test misc_sample_indices_cache_1000_of_1M     ... bench:      61,931 ns/iter (+/- 1,209)
test misc_sample_indices_cache_1000_of_1k     ... bench:      66,852 ns/iter (+/- 1,548)
test misc_sample_indices_cache_100_of_100k    ... bench:       7,502 ns/iter (+/- 487)
test misc_sample_indices_cache_100_of_10k     ... bench:       7,899 ns/iter (+/- 259)
test misc_sample_indices_cache_100_of_1M      ... bench:       7,216 ns/iter (+/- 349)
test misc_sample_indices_cache_100_of_1k      ... bench:       7,218 ns/iter (+/- 211)
test misc_sample_indices_cache_10_of_100k     ... bench:         629 ns/iter (+/- 23)
test misc_sample_indices_cache_10_of_10k      ... bench:         660 ns/iter (+/- 21)
test misc_sample_indices_cache_10_of_1M       ... bench:         594 ns/iter (+/- 27)
test misc_sample_indices_cache_10_of_1k       ... bench:         594 ns/iter (+/- 16)

test misc_sample_indices_floyd_1000_of_100k   ... bench:     118,336 ns/iter (+/- 3,468)
test misc_sample_indices_floyd_1000_of_10k    ... bench:     120,771 ns/iter (+/- 2,713)
test misc_sample_indices_floyd_1000_of_1M     ... bench:     114,288 ns/iter (+/- 4,075)
test misc_sample_indices_floyd_1000_of_1k     ... bench:      69,342 ns/iter (+/- 1,754)
test misc_sample_indices_floyd_100_of_100k    ... bench:       2,376 ns/iter (+/- 95)
test misc_sample_indices_floyd_100_of_10k     ... bench:       2,991 ns/iter (+/- 45)
test misc_sample_indices_floyd_100_of_1M      ... bench:       1,776 ns/iter (+/- 100)
test misc_sample_indices_floyd_100_of_1k      ... bench:       2,011 ns/iter (+/- 69)
test misc_sample_indices_floyd_10_of_100k     ... bench:         127 ns/iter (+/- 4)
test misc_sample_indices_floyd_10_of_10k      ... bench:         166 ns/iter (+/- 10)
test misc_sample_indices_floyd_10_of_1M       ... bench:          96 ns/iter (+/- 4)
test misc_sample_indices_floyd_10_of_1k       ... bench:          95 ns/iter (+/- 4)

test misc_sample_indices_inplace_1000_of_100k ... bench:      37,888 ns/iter (+/- 1,775)
test misc_sample_indices_inplace_1000_of_10k  ... bench:      15,556 ns/iter (+/- 706)
test misc_sample_indices_inplace_1000_of_1M   ... bench:   1,684,064 ns/iter (+/- 32,348)
test misc_sample_indices_inplace_1000_of_1k   ... bench:      10,121 ns/iter (+/- 269)
test misc_sample_indices_inplace_100_of_100k  ... bench:      29,498 ns/iter (+/- 914)
test misc_sample_indices_inplace_100_of_10k   ... bench:       3,368 ns/iter (+/- 141)
test misc_sample_indices_inplace_100_of_1M    ... bench:   1,678,077 ns/iter (+/- 69,038)
test misc_sample_indices_inplace_100_of_1k    ... bench:         813 ns/iter (+/- 33)
test misc_sample_indices_inplace_10_of_100k   ... bench:      28,706 ns/iter (+/- 624)
test misc_sample_indices_inplace_10_of_10k    ... bench:       2,280 ns/iter (+/- 43)
test misc_sample_indices_inplace_10_of_1M     ... bench:   1,661,896 ns/iter (+/- 20,477)
test misc_sample_indices_inplace_10_of_1k     ... bench:         223 ns/iter (+/- 4)

test misc_sample_indices_floyd_hash_1000_of_100k ... bench:      97,342 ns/iter (+/- 1,582)
test misc_sample_indices_floyd_hash_1000_of_10k  ... bench:     102,381 ns/iter (+/- 1,273)
test misc_sample_indices_floyd_hash_1000_of_1M   ... bench:      94,833 ns/iter (+/- 1,037)
test misc_sample_indices_floyd_hash_1000_of_1k   ... bench:     105,910 ns/iter (+/- 1,225)
test misc_sample_indices_floyd_hash_100_of_100k  ... bench:       7,376 ns/iter (+/- 135)
test misc_sample_indices_floyd_hash_100_of_10k   ... bench:       7,726 ns/iter (+/- 281)
test misc_sample_indices_floyd_hash_100_of_1M    ... bench:       7,097 ns/iter (+/- 126)
test misc_sample_indices_floyd_hash_100_of_1k    ... bench:       7,133 ns/iter (+/- 336)
test misc_sample_indices_floyd_hash_10_of_100k   ... bench:         585 ns/iter (+/- 13)
test misc_sample_indices_floyd_hash_10_of_10k    ... bench:         625 ns/iter (+/- 12)
test misc_sample_indices_floyd_hash_10_of_1M     ... bench:         559 ns/iter (+/- 19)
test misc_sample_indices_floyd_hash_10_of_1k     ... bench:         558 ns/iter (+/- 20)

test misc_sample_indices_floyd_tree_1000_of_100k ... bench:     126,655 ns/iter (+/- 2,071)
test misc_sample_indices_floyd_tree_1000_of_10k  ... bench:     130,358 ns/iter (+/- 5,582)
test misc_sample_indices_floyd_tree_1000_of_1M   ... bench:     123,179 ns/iter (+/- 3,541)
test misc_sample_indices_floyd_tree_1000_of_1k   ... bench:     112,819 ns/iter (+/- 3,700)
test misc_sample_indices_floyd_tree_100_of_100k  ... bench:       8,533 ns/iter (+/- 579)
test misc_sample_indices_floyd_tree_100_of_10k   ... bench:       9,081 ns/iter (+/- 542)
test misc_sample_indices_floyd_tree_100_of_1M    ... bench:       8,155 ns/iter (+/- 246)
test misc_sample_indices_floyd_tree_100_of_1k    ... bench:       8,231 ns/iter (+/- 224)
test misc_sample_indices_floyd_tree_10_of_100k   ... bench:         514 ns/iter (+/- 10)
test misc_sample_indices_floyd_tree_10_of_10k    ... bench:         563 ns/iter (+/- 25)
test misc_sample_indices_floyd_tree_10_of_1M     ... bench:         476 ns/iter (+/- 18)
test misc_sample_indices_floyd_tree_10_of_1k     ... bench:         475 ns/iter (+/- 6)
```
Observations:
- performance of `cache` method depends only on `amount` but has high overhead
- performance of Floyd's alg *should* depend only on `amount` but has some variance (it's actually better for large `length`)
- using a `HashSet` or `BTreeSet` to speed up `contains` check in Floyd's (to beat `O(m^2)`) only boosts performance from around 1000 elts, and never beats the `cache` method
```
    // Fastest for various params:
    // amount = 10, length >= 1k: floyd
    // amount = 100, length = 1k: inplace
    // amount = 100, length >= 10k: floyd
    // amount = 1000, length in {1k, 10k, 100k}: inplace
    // amount = 1000, length = 1M: cache
```
I'll do some more testing to try to find the best algorithm (on my system, at least).