@dhardy 2018-06-01
PR updated. Benchmark summary:
```
# old:
test misc_sample_indices_100_of_1G   ... bench:       7,170 ns/iter (+/- 80)
test misc_sample_indices_100_of_1M   ... bench:       7,144 ns/iter (+/- 144)
test misc_sample_indices_100_of_1k   ... bench:         542 ns/iter (+/- 2)
test misc_sample_indices_10_of_1k    ... bench:         593 ns/iter (+/- 6)
test misc_sample_indices_1_of_1k     ... bench:          78 ns/iter (+/- 0)
test misc_sample_indices_1k_of_1G    ... bench:      61,790 ns/iter (+/- 2,031)

# new:
test misc_sample_indices_100_of_1G   ... bench:       1,903 ns/iter (+/- 76)
test misc_sample_indices_100_of_1M   ... bench:       1,857 ns/iter (+/- 87)
test misc_sample_indices_100_of_1k   ... bench:         557 ns/iter (+/- 48)
test misc_sample_indices_10_of_1k    ... bench:         167 ns/iter (+/- 9)
test misc_sample_indices_1_of_1k     ... bench:          26 ns/iter (+/- 0)
test misc_sample_indices_1k_of_1G    ... bench:      61,545 ns/iter (+/- 4,593)
```