@dhardy 2018-06-02
I noticed that we lost one feature: the output is now not fully randomised. Floyd's algorithm has a variant with full randomisation (without requiring extra random draws), but it requires insertions at arbitrary positions. Since this is only used with lengths under 517 anyway, the result isn't too bad, but it does still reduce performance:
```
# before; other tests are above:
test misc_sample_indices_500_of_1G   ... bench:      30,343 ns/iter (+/- 650)
test misc_sample_indices_600_of_1G   ... bench:      35,488 ns/iter (+/- 1,488)
# with full randomisation:
test misc_sample_indices_100_of_1G   ... bench:       2,745 ns/iter (+/- 38)
test misc_sample_indices_100_of_1M   ... bench:       2,732 ns/iter (+/- 100)
test misc_sample_indices_100_of_1k   ... bench:         506 ns/iter (+/- 19)
test misc_sample_indices_10_of_1k    ... bench:         132 ns/iter (+/- 6)
test misc_sample_indices_1_of_1k     ... bench:          26 ns/iter (+/- 1)
test misc_sample_indices_500_of_1G   ... bench:      40,744 ns/iter (+/- 1,988) # uses floyd's alg
test misc_sample_indices_600_of_1G   ... bench:      35,456 ns/iter (+/- 1,232) # uses cached alg
```
Obviously this implies the selection heuristics need adjustment, but is it worth adding up to 77% overhead for a feature not everyone wants? Probably not. On the other hand, some people will want it, and in many cases the overhead is small or even zero.

It is possible to add a `shuffled: bool` parameter to `sample_indices` with very little overhead (except a bit more code in the Floyd impl); is this worthwhile? It also implies that the parameter should be added to `sample_slice`.