@dhardy 2018-06-02
Updated with the mentioned `shuffled` parameter. I don't exactly like the new API, but regarding performance/capabilities this seems to be a good choice. Comments?

New benchmarks:
```
# full shuffling:
test misc_sample_indices_100_of_1G   ... bench:       2,704 ns/iter (+/- 77)
test misc_sample_indices_100_of_1M   ... bench:       2,694 ns/iter (+/- 13)
test misc_sample_indices_100_of_1k   ... bench:         508 ns/iter (+/- 8)
test misc_sample_indices_10_of_1k    ... bench:          71 ns/iter (+/- 0)
test misc_sample_indices_1_of_1k     ... bench:          28 ns/iter (+/- 0)
test misc_sample_indices_400_of_1G   ... bench:      27,075 ns/iter (+/- 177)
test misc_sample_indices_500
# without full shuffling:
test misc_sample_indices_100_of_1G   ... bench:       1,625 ns/iter (+/- 18)
test misc_sample_indices_100_of_1M   ... bench:       1,580 ns/iter (+/- 88)
test misc_sample_indices_100_of_1k   ... bench:         515 ns/iter (+/- 1)
test misc_sample_indices_10_of_1k    ... bench:          75 ns/iter (+/- 0)
test misc_sample_indices_1_of_1k     ... bench:          28 ns/iter (+/- 3)
test misc_sample_indices_400_of_1G   ... bench:      20,399 ns/iter (+/- 116)
test misc_sample_indices_500_of_1G   ... bench:      27,722 ns/iter (+/- 245)
```