@dhardy 2018-06-28
Okay, I finally got around to investigating returning an iterator instead of the type directly. Didn't work. But returning an enum seems to.

Returning variadic iterators is *not* fun. `impl Iterator` doesn't help because it only *hides* the type, and doesn't make it variadic. `Box<Iterator>` would except that the lifetime needs to be tied into the type signature, yet it is omited here (it works for `into_iter()` but not `iter()`). So the only solution seems to be to roll my own type.

I pushed the working changes (messy commits and all) for initial review. Probably I'll squash all those commits. Some things could be renamed, and there is potential for some algorithm tweaks. Also it may be better not to return `impl Iterator` because we don't *need* to update the minimum supported compiler from 1.22.

Benchmarks actually *improved* in some cases and I have *no idea* why (the commit responsible is "temp: abstract over multiple iterator types"). One benchmark got significantly worse (because I decided not to bother with multiple implementations of the `cache` backend, thus it must now return `Vec<usize>` for a test that could use `Vec<u32>`). Some optimisation is possible here but only by adding more backend variants, and I'm not sure its worth it. So, the actual results:
```
# before
running 7 tests
test misc_sample_indices_100_of_1G   ... bench:       1,901 ns/iter (+/- 49)
test misc_sample_indices_100_of_1M   ... bench:       1,863 ns/iter (+/- 73)
test misc_sample_indices_100_of_1k   ... bench:         509 ns/iter (+/- 10)
test misc_sample_indices_10_of_1k    ... bench:          75 ns/iter (+/- 1)
test misc_sample_indices_1_of_1k     ... bench:          28 ns/iter (+/- 0)
test misc_sample_indices_400_of_1G   ... bench:      24,102 ns/iter (+/- 1,463)
test misc_sample_indices_600_of_1G   ... bench:      36,996 ns/iter (+/- 1,584)
# after
running 7 tests
test misc_sample_indices_100_of_1G   ... bench:       1,642 ns/iter (+/- 37)
test misc_sample_indices_100_of_1M   ... bench:       1,587 ns/iter (+/- 100)
test misc_sample_indices_100_of_1k   ... bench:         512 ns/iter (+/- 12)
test misc_sample_indices_10_of_1k    ... bench:          77 ns/iter (+/- 0)
test misc_sample_indices_1_of_1k     ... bench:          32 ns/iter (+/- 1)
test misc_sample_indices_400_of_1G   ... bench:      21,227 ns/iter (+/- 449)
test misc_sample_indices_600_of_1G   ... bench:      42,129 ns/iter (+/- 1,484)
```